Refactoring join container and add organisation
Refactored checkbox changes to only have on main code as there was a lot
of duplicate in the previous version, now I have one function that I am
using for them all.
Changed checkbox values to be id of the item so on the back end i didnt
have to make three api calls to the different tables, reduces number of
request.