Bug - Filter Bug
Find organisations returns all the organisations that exists in the
database for each search option. It does the search independent of
each search option for example if i search with a location and services,
it will return all the organisations that are in that location and all
the organisations that have that services. It is supposed to return all
the organisations in the specified location offering the services
specified. Need to account for all search possibilites.
 Bug remains unfixed.